requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-bootstrap4==22.1
django-ckeditor==6.4.2
django-environ==0.9.0
django-js-asset==2.0.0
django-model-utils==4.2.0
idna==3.3
iniconfig==1.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
testcase==0.1.0
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.10
